refactor: moved helper trace rendering to utils file
